# speed up dispatch on parameterized function types