--- a/src/datatype.c
+++ b/src/datatype.c
@@ -47,6 +47,7 @@ JL_DLLEXPORT jl_methtable_t *jl_new_method_table(jl_sym_t *name, jl_module_t *mo
     mt->kwsorter = NULL;
     mt->backedges = NULL;
     JL_MUTEX_INIT(&mt->writelock);
+    mt->offs = 1;
     return mt;
 }
 
@@ -480,6 +481,8 @@ JL_DLLEXPORT jl_datatype_t *jl_new_datatype(
             if (!abstract) {
                 tn->mt = jl_new_method_table(name, module);
                 jl_gc_wb(tn, tn->mt);
+                if (jl_svec_len(parameters) > 0)
+                    tn->mt->offs = 0;
             }
         }
         t->name = tn;