This should be a safe subset of #21760. Allows splitting on the type of the function for parameterized types. Does not change dispatch on Types (constructors), since it is not strictly better in that case --- many constructors are abstract types that we can't split on. Improving that case will require a more complex change.

Using the benchmark from #21760:
Before:
```
  0.000000 seconds
  0.015294 seconds
  0.001139 seconds (10.00 k allocations: 156.250 KiB)
  0.068189 seconds (489 allocations: 7.641 KiB)
  0.042070 seconds (383 allocations: 26.453 KiB)
  0.043586 seconds (383 allocations: 26.453 KiB)
```
After:
```
  0.000000 seconds
  0.015087 seconds
  0.001219 seconds (10.00 k allocations: 156.250 KiB)
  0.044221 seconds (489 allocations: 7.641 KiB)
  0.000946 seconds (383 allocations: 26.453 KiB)
  0.000969 seconds (383 allocations: 26.453 KiB)
```
(Note we are now able to optimize out the first benchmark completely.)